Adjustments to constructor for keyed SQLCipher databases
The preKeyAction is executed prior to setting the database key, useful for
cipher_default* PRAGMA settings.

The postKeyAction is executed after setting the database key, useful for
non-default (per-connection) PRAGMA settings.

WAL mode will only be force-enabled for non-encrypted databases, to avoid
breaking common SQLCipher useage scenarios. Note that WAL can always be
enabled manually and is automatically persisted so it does not need to be
set every time a database is opened.